Population Line Graph
Added a population line graph.  Currently uses a CSV for reference data, but a JSON probably would work better and I'll try to change over to that.  Displays the y-value of the nearest data point as a tooltip (so you know exactly what the population was in a certain year).