Responsive Design Update
The page should now be responsive.  There is really only one major breakpoint, which decides whether the charts will be displayed in 1 or 2 columns (1 column for mobile, 2 for large tablets or desktop).  The graphs will now also resize themselves to fit the size of the browser window!  Much of the css has also been broken down into a separate stylesheet.  Used CSS Grid to accomplish this (not bootstrap).