Fixed Mobile Devices Not Displaying Properly
#4 bottom & right side of charts are cut off on mobile view

Okay, looked into this further and I think we should be good now.  Turns out that the iframe content was set to certain size by default and would only adjust if the window was resized— this works fine in a browser, but was what was causing the graphs to look weirdly small on tablets and be cut off on phones.  I added a call to adjust to the initial size of the media displayer when the page loads, so now the charts should automatically start out the appropriate size on any device.  Looks good on Chrome's test tools and other third-party test sites I found.