Soft-coded additional stats
Made a big shift to soft-coding the additional stats section. Each stat now has its own unique ID, which is matched in a switch case to a row of data from the master CSV file.  Appending the numeric values (as strings) to each bullet point is now done functionally, with fewer extraneous cases.  There is now one boolean flag parameter (true by default) which allows the rank of the statistic to be shown when true.  Error handling is implemented for several cases:
  - A new statistic has been added, but a switch case for it has not yet been implemented.
  - The switch case could not match the element ID to any metrics in the master CSV (sometimes the result of a misspelling)
  - A switch case exists and matches an existing metric, but the data for that metric has not been defined

Note that this implementation now includes a line of jQuery, which has been sourced into the file from Google's online library host.

Also fixed a bug that would occassionally display improper suffixes for the rank of a statistic (such as 11st).

Also changed the update method to reflect previous changes.  When the selected county is changed, all that needs to be done is to update the available dataset and then make a call to the main switch case control function.

Also made slight documentation improvements.

Finally, reworked some CSS to make the additional stats section appear hopefully nicer on most devices.  This included adding in 3 or 4 new breakpoints which the stats section will adapt to depending on the width of the view media.